Move chainLogos to constants
