--- src/app/components/Multistaking/chainLogos.ts
@@ -1,15 +1,0 @@
-import babylon from "@/app/assets/chains/babylon-genesis.png";
-import cosmos from "@/app/assets/chains/cosmos.png";
-import ethereum from "@/app/assets/chains/ethereum.png";
-import placeholder from "@/app/assets/chains/placeholder.svg";
-import sui from "@/app/assets/chains/sui.png";
-
-export const chainLogos = {
-  babylon,
-  cosmos,
-  ethereum,
-  sui,
-  placeholder,
-} as const;
-
-export { babylon, cosmos, ethereum, sui, placeholder };
